sdialectic-dspy/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sdialectic-core"
version = "2.0.0"  # Or whatever version you want
dependencies = [
    "fastapi>=0.110.0",  # Use a more recent version
    "pydantic>=2.7.0",  # Use a more recent version
    "matplotlib>=3.8.0",
    "pandas>=2.2.0",
    "datasets>=2.17.0",
    "bioc>=2.1.0",
    "dspy",
    "google-genai",
    "uvicorn[standard]>=0.30.0",
]

[tool.setuptools.packages.find]
include = ["core", "dspy_modules", "optimization"]  # Explicitly list your packages
exclude = ["tests*"]  # Optional: Exclude test directories